requirements.txt
flair==0.4.5
matplotlib==3.1.3
nltk==3.4.5
numpy==1.18.1
pandas==1.0.0
rouge==1.0.0
pylanguagetool==0.9.0
pyyaml==5.3.1
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
sentence-transformers==0.2.6.1
statsmodels==0.11.1
torch==1.4.0
tqdm==4.42.1
xlrd==1.2.0